requirements.txt
Flask
PyJWT
python-dotenv
SQLAlchemy
mysql-connector-python
werkzeug
